chore: consolidate converter/postgres test files by module (21 -> 8)
internal/converter/postgres/ mixed 11 source files with 21 test files,
making the directory hard to scan. Go requires tests to live next to
the package they test, so consolidation (not relocation) is the only
toolchain-legal cleanup:

- sync_tableddl_test.go   += sync_tableddl_p2, sync_ddl_batch,
                             sync_unsigned_decimal, sync_timestamp
- sync_viewddl_test.go    += sync_batch_e, view_function_filter,
                             sync_literal_protect
- sync_functions_test.go  += function_control, function_decl
- sync_privilege_test.go  := sync_privilege_p0 + sync_privilege_batch
- sync_data_test.go       += sync_data_p2, sync_validate
- manager_test.go         := manager_ctx + manager_progress
- sync_indexes_test.go / sync_sequence_test.go kept as-is

Pure mechanical merge: package-scope identifiers were already unique
across the 21 files (they compiled together), so no renames or edits
to test bodies were needed — only package/import boilerplate removed.
Verified: 153 Test + 2 Benchmark functions before and after; gofmt
clean; full suite green under -race
